Rework logic

+++ Constants.cs
@@ -1,7 +1,8 @@
-namespace orbital_mechanics{
+namespace orbital_mechanics
+{
     public static class Constants
     {
-        public static double Gravitational_Constant = 6.674e-11;
+        public static float Gravitational_Constant = 6.674e-11f;
 
     }
 }
